works/python/requirements.txt
ipaddress==1.0.23
Flask==2.3.3
requests==2.31.0
langchain==0.0.350
openai==1.3.0
python-dotenv==1.0.0
chromadb==0.4.15
beautifulsoup4==4.12.2
newspaper3k==0.2.8
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0 